Complete confirmation modal

+++ KayakCove.Infrastructure/Repositories/CategoryRepository.cs
@@ -36,10 +36,13 @@ public class CategoryRepository : ICategoryRepository
         await _context.SaveChangesAsync();
     }
 
-    public async Task DeleteCategoryAsync(int id)
+    public async Task<bool> DeleteCategoryAsync(int id)
     {
         var category = await _context.Categories.FindAsync(id);
+        if (category is null) return false;
+        
         _context.Categories.Remove(category);
         await _context.SaveChangesAsync();
+        return true;
     }
 }
